requirement.txt
pandas
opencv-python
seaborn
os-sys
scikit-learn
pyfiglet
statsmodel
clint
matplotlib